Fix some references, add a finalize create service test, move PATCH contact tests into their own file
