+++ packages/twilio-client/mockClient/taskRouterWorkspaces.ts
@@ -1,0 +1,5 @@
+export let taskRouterWorkspaces: any[] = [];
+
+export const setMockTaskRouterWorkspaces = (workspaces: any[]) => {
+  taskRouterWorkspaces = workspaces;
+};
